+++ App.js
@@ -3,17 +3,17 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { Provider } from 'react-redux';
+import Chat from './src/pages/chat';
+import ChatList from './src/pages/chatList';
 import demo from './src/pages/demo';
 import s3 from './src/pages/s3'
 import home from './src/pages/home';
 import login from './src/pages/login';
 import map from './src/pages/Map';
-import register from './src/pages/register';
-import Chat from './src/pages/chat';
-import ChatList from './src/pages/chatList';
+import EventDisplay from './src/pages/Map/components/eventPage/eventDisplay';
 import EventPage from './src/pages/Map/components/eventPage/eventPage';
-import Profile from './src/pages/profile'
-import EventDisplay from './src/pages/Map/components/eventPage/eventDisplay'
+import Profile from './src/pages/profile';
+import register from './src/pages/register';
 import { store } from './src/store';
 
 const Stack = createNativeStackNavigator();
@@ -21,7 +21,7 @@ export default function App() {
   return (
     <Provider store={store}>
       <NavigationContainer>
-        <Stack.Navigator initialRouteName="Login">
+        <Stack.Navigator initialRouteName="Map">
           <Stack.Screen
             name="Map"
             component={map}