+++ src/pages/login/index.js
@@ -2,7 +2,7 @@ import {
   getAuth,
   signInWithEmailAndPassword
 } from 'firebase/auth';
-import React, { useState, useSelector } from 'react';
+import React, { useState } from 'react';
 import {
   KeyboardAvoidingView,
   StyleSheet,
@@ -11,9 +11,9 @@ import {
   TouchableOpacity,
   View
 } from 'react-native';
-import firebaseConfig from '../../../authBase';
-import { fetchUser } from '../../services/api'
 import { useDispatch } from 'react-redux';
+import firebaseConfig from '../../../authBase';
+import { fetchUser } from '../../services/api';
 
 function LoginScreen({ navigation }) {
   const dispatch = useDispatch()
@@ -25,7 +25,7 @@ function LoginScreen({ navigation }) {
     signInWithEmailAndPassword(auth, email, password)
       .then((userCredentials) => {
         const { user } = userCredentials;
-        navigation.replace('Profile');
+        navigation.replace('Map');
         console.log('Logged in with:', user.email);
         dispatch(fetchUser(user.email))
       })