+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.1.3] - 2022-10-29
+## Added
+- added navigation function to navigation bar
+- added create event dialog
+
 # [1.1.2] - 2022-10-29
 ## Added
 - show different event display for different roles(host, participant), can navigate to event hosted or participanted through profile